Fix nested archives in CI artifacts
`actions/upload-artifact@v2` zips the given path, which is already a zip,
resulting in nested zips in the artifacts. We can have `upload-artifact`
run on the results instead, which requires moving the cleaning to the
next build and placing the uploads in-between as each build was removing
its output aside fromits created zip.